feat: initial support for ssh auth, refactorings
