+++ internal/configuration/providerconfig.go
@@ -5,19 +5,10 @@
 package configuration
 
 import (
-	"context"
 	"fmt"
-	"os"
-	"path/filepath"
+	"itiquette/git-provider-sync/internal/model"
 	"strings"
 
-	"itiquette/git-provider-sync/internal/model"
-
-	"github.com/knadh/koanf/parsers/dotenv"
-	"github.com/knadh/koanf/parsers/yaml"
-	"github.com/knadh/koanf/providers/env"
-	"github.com/knadh/koanf/providers/file"
-	"github.com/knadh/koanf/v2"
 	"github.com/rs/zerolog"
 )
 
@@ -28,6 +19,7 @@ type ProviderConfig struct {
 	Token            string            `koanf:"token"`
 	User             string            `koanf:"user"`
 	Group            string            `koanf:"group"`
+	GitInfo          model.GitInfo     `koanf:"gitoption"`
 	Exclude          map[string]string `koanf:"exclude"`
 	Include          map[string]string `koanf:"include"`
 	Providerspecific map[string]string `koanf:"providerspecific"`
@@ -35,8 +27,8 @@ type ProviderConfig struct {
 
 // String returns a string representation of ProviderConfig, masking the token.
 func (p ProviderConfig) String() string {
-	return fmt.Sprintf("ProviderConfig: Provider: %s, Domain: %s, Token: <****>, User: %s, Group: %s, Exclude: %v, Include: %v, ProviderSpecific: %v",
-		p.Provider, p.Domain, p.User, p.Group, p.Exclude, p.Include, p.Providerspecific)
+	return fmt.Sprintf("ProviderConfig: Provider: %s, Domain: %s, Token: <****>, User: %s, GitInfo: %v,  Group: %s, Exclude: %v, Include: %v, ProviderSpecific: %v",
+		p.Provider, p.Domain, p.User, p.GitInfo, p.Group, p.Exclude, p.Include, p.Providerspecific)
 }
 
 // DebugLog logs the ProviderConfig details at debug level.
@@ -104,129 +96,4 @@ func (a AppConfiguration) DebugLog(logger *zerolog.Logger) {
 			target.DebugLog(logger).Msg(key + ":TargetProvider")
 		}
 	}
-}
-
-// configuration/configuration.go
-
-type ConfigLoader interface {
-	LoadConfiguration(ctx context.Context) (*AppConfiguration, error)
-}
-
-type DefaultConfigLoader struct{}
-
-// LoadConfiguration loads the application configuration from various sources.
-func (DefaultConfigLoader) LoadConfiguration(ctx context.Context) (*AppConfiguration, error) {
-	cliOption := model.CLIOptions(ctx)
-	appConfig := &AppConfiguration{}
-
-	if err := ReadConfigurationFile(appConfig, cliOption.ConfigFilePath, cliOption.ConfigFileOnly); err != nil {
-		return nil, fmt.Errorf("failed to read configuration file: %w", err)
-	}
-
-	for _, config := range appConfig.Configurations {
-		if err := ValidateConfiguration(config); err != nil {
-			return nil, fmt.Errorf("failed to validate configuration: %w", err)
-		}
-	}
-
-	return appConfig, nil
-}
-
-func ReadConfigurationFile(appConfiguration *AppConfiguration, configfile string, configfileOnly bool) error {
-	const (
-		xdgConfigHomeEnv        = "XDG_CONFIG_HOME"
-		xdgConfigHomeConfigPath = "/gitprovidersync/" + "gitprovidersync.yaml"
-		dotEnvFilename          = ".env"
-	)
-
-	koanfConf := koanf.New(".")
-	xdgConfigfileExists, xdgConfigFilePath := hasXDGConfigFile(xdgConfigHomeEnv, xdgConfigHomeConfigPath)
-	localConfigfileExists := hasLocalConfigFile(configfile)
-	dotEnvFileExists, dotEnvFilePath := hasDotEnvFile(dotEnvFilename)
-
-	// xdg config file
-	if xdgConfigfileExists && !configfileOnly {
-		if err := koanfConf.Load(file.Provider(xdgConfigFilePath), yaml.Parser()); err != nil {
-			return fmt.Errorf("error loading xdg_config_home configuration. %w", err)
-		}
-	}
-
-	// local config file
-	if localConfigfileExists {
-		if err := koanfConf.Load(file.Provider(configfile), yaml.Parser()); err != nil {
-			return fmt.Errorf("error loading config: %w", err)
-		}
-	}
-
-	// .env file
-	if dotEnvFileExists && !configfileOnly {
-		err := koanfConf.Load(file.Provider(dotEnvFilePath), dotenv.ParserEnv("", "_", strings.ToLower))
-		if err != nil {
-			return fmt.Errorf("error loading dotenvfile config: %w", err)
-		}
-	}
-	// env variable (prefix: GPS_)
-	if !configfileOnly {
-		if err := koanfConf.Load(env.Provider("GPS_", ".", func(s string) string {
-			return strings.ReplaceAll(strings.ToLower(
-				strings.TrimPrefix(s, "GPS_")), "_", ".")
-		}), nil); err != nil {
-			return fmt.Errorf("failed to read environment conf: %w", err)
-		}
-	}
-
-	// Unmarshal the YAML data into the config struct
-	if err := koanfConf.Unmarshal("", appConfiguration); err != nil {
-		panic(fmt.Errorf("error unmarshalling yaml config: %w", err))
-	}
-
-	if len(appConfiguration.Configurations) == 0 {
-		panic("No configuration could be found!")
-	}
-
-	return nil
-}
-
-func hasXDGConfigFile(xdgconfighome string, xdgconfighomeconfigpath string) (bool, string) {
-	xdgConfigfileExists := false
-
-	var xdgConfigFilePath string
-
-	envValue, xdgHomeIsSet := os.LookupEnv(xdgconfighome)
-	if xdgHomeIsSet {
-		xdgConfigFilePath = filepath.Join(envValue, xdgconfighomeconfigpath)
-		if _, err := os.Stat(xdgConfigFilePath); err == nil {
-			xdgConfigfileExists = true
-		}
-	}
-
-	return xdgConfigfileExists, xdgConfigFilePath
-}
-
-func hasLocalConfigFile(configFile string) bool {
-	localConfigfileExists := false
-	if _, err := os.Stat(configFile); err == nil {
-		localConfigfileExists = true
-	}
-
-	return localConfigfileExists
-}
-
-func hasDotEnvFile(dotEnvFilePath string) (bool, string) {
-	dotEnvFileExists := false
-
-	envValue, testConfigHomeIsSet := os.LookupEnv("GPS_TESTCONFIG_HOME")
-	if testConfigHomeIsSet {
-		dotEnvFilePath = filepath.Join(envValue, dotEnvFilePath)
-	}
-
-	if _, err := os.Stat(dotEnvFilePath); err == nil {
-		dotEnvFileExists = true
-	}
-
-	return dotEnvFileExists, dotEnvFilePath
-}
-
-func splitAndTrim(s string) []string {
-	return strings.Split(strings.ReplaceAll(s, " ", ""), ",")
 }
